Time: 4 ms (21.25%), Space: 5.8 MB (97.17%) - LeetHub

+++ 1137-n-th-tribonacci-number/1137-n-th-tribonacci-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int tribonacci(int n) {
-        long long int ans=0;
+        int ans=0;
         int a=0,b=1,c=1;
         
         if(n==0 || n==1)
